refactor: extract gitlab.types.ts and fix import shadowing (#319)
Move shared types (GitlabUser, GitlabRepositoryFull, GitlabRepository,
GitlabBranch, GitlabFile, GitlabTreeItem, ConflictResolution,
SyncOperation, Transaction) into dedicated gitlab.types.ts module.
Fix variable shadowing: rename local gitlabRepositories -> repos in
listGitlabRepositories.
